+++ weapon_tweaker/scripts/mods/weapon_tweaker/weapon_tweaker_data.lua
@@ -1304,6 +1304,7 @@ local data = {
                     { setting_id = "wt_cog_hammer_heavy_speed_nerf", type = "checkbox", default_value = false },
                     { setting_id = "wt_dual_axes_cleave", type = "checkbox", default_value = true },
                     { setting_id = "wt_dual_axes_light_crit", type = "checkbox", default_value = true },
+                    { setting_id = "wt_executioner_light_headshot_bonus", type = "checkbox", default_value = false },
                     { setting_id = "wt_greataxe_light_crit", type = "checkbox", default_value = true },
                     { setting_id = "wt_mace_sword_speed_nerf", type = "checkbox", default_value = false },
                     { setting_id = "wt_bolt_staff_primary_overcharge_reduction", type = "checkbox", default_value = false },